+++ Command/List.hs
@@ -1,0 +1,10 @@
+module Command.List where
+
+import qualified Git as Git
+
+
+runList :: IO ()
+runList = do
+    branches <- Git.getBranches
+    putStrLn $ unlines branches
+    return ()